Reorganize some code
